rename StoreIdentityEmailRequest file
