+++ app/Http/Requests/Api/Identity/Emails/StoreIdentityEmailRequest.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace App\Http\Requests\Api\Identity;
+namespace App\Http\Requests\Api\Identity\Emails;
 
 use App\Rules\IdentityEmailUniqueRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class IdentityEmailStoreRequest extends FormRequest
+class StoreIdentityEmailRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
      *
      * @return bool
      */
-    public function authorize()
+    public function authorize(): bool
     {
         return true;
     }
@@ -22,7 +22,7 @@ class IdentityEmailStoreRequest extends FormRequest
      *
      * @return array
      */
-    public function rules()
+    public function rules(): array
     {
         return [
             'email' => [
